Bed and Bookshelf
Made 3 voicelines for Bed and 3 for bookshelf. Edit were made to take out background sound of the Heater noise. Used a blanket to keep the echo from happening when recording. Noise reduction for the heater noise in the background. Covered headphone mic with blanket and spoke into the blanket to keep unneccessary plosives from happening.