Audio Change
Added a Laptop mixer and made the laptop sound more like it was playing from a speaker rather than sounding like there is a person there. Changed the location of the Door VO to be a bit closer, before it was a bit farther making it hard to hear. Changed the Bed and book Volume a bit down, and changed the volume on music. Music was a bit quiet and raised it just slightly.